WIP make controller behave like an IDP
