+++ core/auth/fake/interfaces/controller.go
@@ -1,8 +1,9 @@
-package fake
+package interfaces
 
 import (
 	"bytes"
 	"context"
+	"errors"
 	"html/template"
 	"net/http"
 
@@ -18,13 +19,18 @@ type (
 
 	viewData struct {
 		FormURL string
+		Message string
 	}
 )
+
+const errMissingUsername = "missing username"
+const errInvalidUser = "invalid user"
 
 const defaultIDPTemplate = `
 <body>
   <h1>Login!</h1>
   <form name="fake-idp-form" action="{{.FormURL}}" method="post">
+	<div>{{.Message}}</div>
 	<label for="username">Username</label>   
 	<input type="text" name="username" id="username">
 	<label for="password">Password</label>
@@ -55,10 +61,20 @@ func (c *idpController) Inject(
 }
 
 // Auth action to simulate OIDC / Oauth Login Page
-func (c *idpController) Auth(_ context.Context, r *web.Request) web.Result {
+func (c *idpController) Auth(ctx context.Context, r *web.Request) web.Result {
 	broker, err := r.Query1("broker")
 	if err != nil || broker == "" {
 		return c.responder.ServerError(err)
+	}
+
+	formError := errors.New("")
+
+	if postValues, err := r.FormAll(); len(postValues) > 0 && err != nil {
+		formError = c.handlePostValues(ctx, postValues, broker)
+
+		if formError == nil {
+			return c.responder.RouteRedirect("core.auth.callback(broker)", map[string]string{"broker": broker})
+		}
 	}
 
 	// get formURL to callback with broker filled in
@@ -70,6 +86,7 @@ func (c *idpController) Auth(_ context.Context, r *web.Request) web.Result {
 	if c.template != "" {
 		return c.responder.Render(c.template, viewData{
 			FormURL: formURL.String(),
+			Message: formError.Error(),
 		})
 	}
 
@@ -88,6 +105,7 @@ func (c *idpController) Auth(_ context.Context, r *web.Request) web.Result {
 		body,
 		viewData{
 			FormURL: formURL.String(),
+			Message: formError.Error(),
 		})
 	if err != nil {
 		return c.responder.ServerError(err)
@@ -98,4 +116,16 @@ func (c *idpController) Auth(_ context.Context, r *web.Request) web.Result {
 		Status: http.StatusOK,
 		Body:   body,
 	}
+}
+
+func (c *idpController) handlePostValues(ctx context.Context, values map[string][]string, broker string) error {
+	// TODO: make this configurable
+
+	if _, ok := values["username"]; !ok {
+		return errors.New(errMissingUsername)
+	}
+
+	// TODO: check user against config and save in session
+
+	return nil
 }
